refactors code and remove unnessasary files
